Finished the bullet, i will add a bullets array in the player class (for collision). I also added a dVector2d class, its the same as vector2d but for doubles. need that for the bullet class calculations

+++ source/attackHand/attackHand.cpp
@@ -44,9 +44,9 @@ void attackHand::tick(double deltaTime)
             attackSlowDown = 3.f;
         }
         if(playerObj->flip){
-            screenLocation = playerObj->screenLocation + vector2d(attackOffset, (int)(playerObj->playerSize.Y/2.7));
+            screenLocation = playerObj->screenLocation + vector2d(attackOffset, playerObj->playerSize.Y/2.7);
         }else{
-            screenLocation = playerObj->screenLocation + vector2d(-1*attackOffset, (int)(playerObj->playerSize.Y/2.7));
+            screenLocation = playerObj->screenLocation + vector2d(-1*attackOffset, playerObj->playerSize.Y/2.7);
         }
     }
 }
